Fixes #2. Set a default editor.
This commit adds a new pseudo-language id which can be used to set a
default editor to be used in the absence of a configuration for the
current language id.

It also adds a context menu entry to launch the command.

In the absence of any language configuration, the default editor of the
OS will be used.

Example configuration:
```
"vscode-opn.perLang": {
    "opnOptions": [
	{
	    "forLang": "html",
	    "openInApp": "chrome",
	},
	{
	    "forLang": "default",
	    "openInApp": "notepad",
	    "isUseFsPath": true
	}
    ]
}
```