Fixes #2 and regression in previous commit.
This commit fixes the wrong behaviour of opening the active file instead
of the file from which the context-menu action was invoked.

Now this only passes a `TextDocument` instead of a complete `TextEditor`
to other functions (mainly in `fileService.js`) to help reduce
complexity and allow code-reuse.

This also rolled back the change in `jsconfig.json`.